src/requirements.txt
attrs==19.3.0
more-itertools==8.2.0
packaging==20.3
peewee==3.13.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
python-chess==0.30.1
six==1.14.0
wcwidth==0.1.9
click==7.1.2
